commit n sinc

+++ tennis-game/SetHistoryText/SetHistoryText.cpp
@@ -10,7 +10,7 @@ std::string SetHistoryText::getSetHistoryText( int player ) {
 std::string SetHistoryText::_buildString( int player ) {
     std::map< int, int > setHistory = player == 1 ? _gameState->getPlayer1SetHistory() : _gameState->getPlayer2SetHistory();
     std::string setHistoryText = ""; 
-    for ( int i = 0; i < setHistory.size(); i++ ) {
+    for ( long unsigned int i = 0; i < setHistory.size(); i++ ) {
         setHistoryText += std::to_string( setHistory[ i ] );
         setHistoryText += " "; }
     return setHistoryText; }
